requirements.txt
requests
openai
pandas
json
openpyxl
pyyaml
os